+++ src/main/java/com/rcoem/filmrentalstore/repository/FilmRepository.java
@@ -3,7 +3,7 @@ package com.rcoem.filmrentalstore.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.rcoem.filmrentalstore.entities.Film;
 
-public interface FilmRepository extends JpaRepository<Film, Long> {
+public interface FilmRepository extends JpaRepository<Film, Short> {
     
 }
